requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
black==23.10.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.0
fastapi-pagination==0.12.12
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
